Добавила или подредактировала бэк и фронт

+++ src/main/java/ru/praktika/kotouslugi/dao/KittensRepository.java
@@ -4,5 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import ru.praktika.kotouslugi.model.Kittens;
 import ru.praktika.kotouslugi.model.Requisition;
 
-public interface KittensRepository extends CrudRepository<Kittens, Integer> {
+
+//ВСЕ ПЛОХО ЭТО УДАЛИТЬ!!
+public interface KittensRepository extends CrudRepository<Kittens, Long> {
 }
